+++ spec/lib/dawn/codesake_knowledgebase_spec.rb
@@ -1034,4 +1034,9 @@ describe "The Codesake Dawn knowledge base" do
     sc.should_not   be_nil
     sc.class.should == Dawn::Kb::CVE_2014_7818
   end
+  it "must have test for OSVDB_120415" do
+    sc = kb.find("OSVDB_120415")
+    sc.should_not   be_nil
+    sc.class.should == Dawn::Kb::OSVDB_120415
+  end
 end